Treat hyphens as word parts for motion, completion
In Raku variables and sub/method names use "kebab case". For example

    my $user-id = 123;
    my $success = verify-admin($user-id);

Adding '-' to the `extra_word_chars` option list (which is only `_`
by default) would let `user-id` or `verify-admin` appear in the code
completion list, and allow selecting them whole with a motion like "w".

There might be something more sophisticated we can do here, but this
feels like an improvement on the status quo.